Clean up compiled simulation interfaces
