+++ docs/source/binaryInterface.rst
@@ -58,3 +58,19 @@ Arguments
    Path for the result series.  Results are written as
    ``core_result_phi_ase``, ``core_result_mse``, ``core_result_total_rays``,
    and ``core_result_dndt_ase`` records.
+
+Compiled Simulation Mode
+------------------------
+
+``Simulation.runSteps(...)`` uses ``calcPhiASE --run-simulation`` rather than
+running a Python pump loop. Python writes the initial geometry, material,
+spectra, beta state, and run-control attributes; C++/Alpaka advances the
+requested steps and writes a snapshot after each one. The snapshots contain the
+updated beta fields, ASE results, and pump and ASE derivatives. The first
+snapshot also carries the static context needed to read the series on its own.
+
+The run control selects a time step, step count, pump-step limit, and one of the
+compiled integrators (explicit Euler, Heun, midpoint, RK4, implicit Euler, or
+exponential Euler). The supported pump routine is
+``one-dimensional-z-traversal``. It is configured through ``PumpProperties``;
+custom Python pump routines are not part of this execution path.
